fix: All-view UX polish (sticky collapse, tag row, date labels, alignment) (#57)
* fix: All-view UX polish (sticky collapse, tag row, date labels, alignment)

Nine small UX fixes on the Tasks All-view: sticky high-priority collapse
state across navigation, live counts on High Priority zones, tidier
filter-chip row (no "Tags" label, right-aligned Clear filters), search box
moved to the top of every screen, a reserved 2-row tag slot so kanban
columns stop jumping between boards, abbreviated weekday with the date on
its own line under each column header, and alphabetized tags in the
Task List quick-edit.


* docs(prd): update PRODUCT_REQUIREMENTS_DOCUMENT.MD for PR #57 [skip deploy]

* docs(arch): update ARCHITECTURE.MD for PR #57 [skip deploy]

---------

+++ frontend/src/utils/taskDateUtils.ts
@@ -22,7 +22,7 @@ export function formatDateWithDay(dateStr: string): string {
   const d = new Date(dateStr + 'T00:00:00');
   const month = d.toLocaleDateString(undefined, { month: 'long' });
   const day = d.getDate();
-  const dayName = d.toLocaleDateString(undefined, { weekday: 'long' });
+  const dayName = d.toLocaleDateString(undefined, { weekday: 'short' });
   return `${month} ${day}, ${dayName}`;
 }
